part of the way to favicon perfection
the problem at the moment is that the android-chrome favicons
are called from site.webmanifest, which then has incorrect paths
for those assets. A quick attempt to make this an erb file using
asset_path tags didn't work out, so this will need more investigation.

Keep in mind that if we make an upgrade task for Orchid, we will want to move
the user's favicon.png file to favicons/ in the future